pyproject.toml
[tool.poetry]
name = "ml-pipeline"
version = "0.1.0"
description = ""
authors = ["Tsuyoshi Tsukahara <t.tsukaxara@gmai.com>"]
readme = "README.md"
packages = [{include = "ml_pipeline"}]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.1"
pandas = "^1.5.2"
pandera = "^0.13.4"
pydantic = "^1.10.4"
prefect = "^2.7.5"
hydra-core = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
